Backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
surya-ocr>=0.4.0
pdf2image>=1.17.0